Add examples of calculations in a google map
Added custom property value, and did some summing based on if  a marker
is captured in the drawn polygon.  Also, display the area of the polygon.
This could be useful as an example for showing customers on a map, and 
how much they spend on average, grouped by a drawn on shape.